Fix styling of zendesk component
